+++ tests/test_engines.py
@@ -1,3 +1,5 @@
+import warnings
+
 from . import mock, BaseRedmineTestCase, Redmine
 
 from redminelib import engines, exceptions
@@ -91,9 +93,16 @@ class BaseEngineTestCase(BaseRedmineTestCase):
         self.assertRaises(exceptions.UnknownError, lambda: self.redmine.engine.request('get', self.url))
 
     def test_http_protocol_exception(self):
-        self.response.history = [mock.Mock()]
-        self.redmine.url = 'http://foo.bar'
+        self.response.history = [mock.Mock(**{'status_code': 301, 'request.url': 'http://foo.bar'})]
         self.assertRaises(exceptions.HTTPProtocolError, lambda: self.redmine.engine.request('get', self.url))
+
+    def test_redirect_warning(self):
+        self.response.history = [mock.Mock(**{'status_code': 301, 'request.url': 'https://www.foo.bar'})]
+        with warnings.catch_warnings(record=True) as w:
+            warnings.simplefilter('always')
+            self.redmine.engine.request('get', self.url)
+            self.assertEquals(len(w), 1)
+            self.assertIs(w[0].category, exceptions.PerformanceWarning)
 
     def test_engine_is_picklable(self):
         import pickle